requirements.txt
pywebview>=4.0
nbtlib>=1.0
pyinstaller>=5.0 